password changing done

+++ updatePassword.php
@@ -6,26 +6,36 @@ include("connect.php");
         $pass2 = $_POST['pass2'];
         $email = $_POST['email'];
         $token = $_POST["t"];
-
-        //Delete the token form database
-        $get_mail_query = "DELETE from resetpassword where token='$token'";
-        mysqli_query($connect, $get_mail_query);
-
-        //get user id to update password in login table
-        $get_id = "SELECT user_id from user where email='$email'";
-        $id = mysqli_query($connect, $get_id);
-        if(mysqli_num_rows($id)>0){
-            if($row = mysqli_fetch_array($id)){
-                $uid = $row['user_id'];
-                $update_password="UPDATE login set password='$pass1' where user_id='$uid'";
-                if(mysqli_query($connect, $update_password)){
-                    echo "<script>alert('reset successfull');</script>";
-                    echo '<script>window.location.href = "login.html";</script>';
-                }
-            }
-        }else{
-            echo "error in fetching user_id";
+        if($pass1 !=$pass2)
+        {
+            echo '<script language="javascript">';
+            echo 'alert("Password Does Not Match");';
+            echo 'window.history.back();';
+            echo '</script>';
         }
+        else
+        {
+        //Delete the token form database
+                $get_mail_query = "DELETE from resetpassword where token='$token'";
+                mysqli_query($connect, $get_mail_query);
+
+                //get user id to update password in login table
+                $get_id = "SELECT user_id from user where email='$email'";
+                $id = mysqli_query($connect, $get_id);
+                if(mysqli_num_rows($id)>0){
+                    if($row = mysqli_fetch_array($id)){
+                        $uid = $row['user_id'];
+                        $update_password="UPDATE login set password='$pass1' where user_id='$uid'";
+                        if(mysqli_query($connect, $update_password)){
+                            echo "<script>alert('Reset successfull');</script>";
+                            echo '<script>window.location.href = "login.html";</script>';
+                        }
+                    }
+                }else{
+                    echo "Error in fetching user_id";
+                }
+        }
+        
     }
     else{
         echo "<script>alert('wrong details entered');</script>";
